+++ pkg/webui/components/table/cell/index.js
@@ -24,7 +24,7 @@ import style from './cell.styl'
 const Cell = ({
   className,
   component: Component,
-  centered,
+  align,
   small,
   colSpan,
   width,
@@ -32,7 +32,9 @@ const Cell = ({
   ...rest
 }) => {
   const cellClassNames = classnames(className, style.cell, {
-    [style.cellCentered]: centered,
+    [style.cellCentered]: align === 'center',
+    [style.cellLeft]: align === 'left',
+    [style.cellRight]: align === 'right',
     [style.cellSmall]: small,
   })
 
@@ -46,8 +48,8 @@ const Cell = ({
 }
 
 Cell.propTypes = {
-  /** A flag indicating whether the data in the cell should be centered. */
-  centered: PropTypes.bool,
+  /** A flag indicating how the text in the cell should be aligned. */
+  align: PropTypes.oneOf(['left', 'right', 'center']),
   children: PropTypes.node,
   className: PropTypes.string,
   /** The number of columns that the cell should occupy. */
@@ -61,7 +63,7 @@ Cell.propTypes = {
 }
 
 Cell.defaultProps = {
-  centered: false,
+  align: undefined,
   children: undefined,
   className: undefined,
   colSpan: 1,