+++ pkg/webui/console/lib/feature-checks.js
@@ -221,15 +221,9 @@ export const mayPerformAdminActions = {
   check: isAdmin => isAdmin,
 }
 
-export const mayManageUsers = {
-  rightsSelector: selectUserIsAdmin,
-  check: mayPerformAdminActions.check,
-}
-
-export const mayPurgeEntities = {
-  rightsSelector: selectUserIsAdmin,
-  check: mayPerformAdminActions.check,
-}
+export const mayManageUsers = mayPerformAdminActions
+export const mayPurgeEntities = mayPerformAdminActions
+export const mayConfigurePacketBroker = mayPerformAdminActions
 
 // Composite feature checks.
 export const mayViewApplications = {