+++ pkg/webui/console/store/reducers/index.js
@@ -23,6 +23,7 @@ import {
   getCombinedDeviceId,
   getApiKeyId,
   getCollaboratorId,
+  getPacketBrokerNetworkId,
 } from '@ttn-lw/lib/selectors/id'
 import { createNamedPaginationReducer } from '@ttn-lw/lib/store/reducers/pagination'
 import fetching from '@ttn-lw/lib/store/reducers/ui/fetching'
@@ -36,6 +37,7 @@ import { SHARED_NAME as DEVICE_SHARED_NAME } from '@console/store/actions/device
 import { SHARED_NAME as USER_SHARED_NAME } from '@console/store/actions/users'
 import { SHARED_NAME as API_KEYS_SHARED_NAME } from '@console/store/actions/api-keys'
 import { SHARED_NAME as COLLABORATORS_SHARED_NAME } from '@console/store/actions/collaborators'
+import { SHARED_NAME as PACKET_BROKER_NETWORKS_SHARED_NAME } from '@console/store/actions/packet-broker'
 
 import user from './user'
 import users from './users'
@@ -61,6 +63,7 @@ import js from './join-server'
 import gatewayStatus from './gateway-status'
 import is from './identity-server'
 import deviceRepository from './device-repository'
+import packetBroker from './packet-broker'
 
 export default history =>
   combineReducers({
@@ -107,10 +110,15 @@ export default history =>
       gateways: createNamedPaginationReducer(GATEWAY_SHARED_NAME, getGatewayId),
       organizations: createNamedPaginationReducer(ORGANIZATION_SHARED_NAME, getOrganizationId),
       users: createNamedPaginationReducer(USER_SHARED_NAME, getUserId),
+      packetBrokerNetworks: createNamedPaginationReducer(
+        PACKET_BROKER_NETWORKS_SHARED_NAME,
+        getPacketBrokerNetworkId,
+      ),
     }),
     router: connectRouter(history),
     js,
     gatewayStatus,
     is,
     deviceRepository,
+    packetBroker,
   })